toollama/API/requirements.txt
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.8.0
archiveis==0.0.9
Authlib==1.4.1
beautifulsoup4==4.13.3
blinker==1.9.0
brotli==1.1.0
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cozepy==0.12.0
cryptography==44.0.1
ddg==0.2.2
distro==1.9.0
duckduckgo_search==7.5.0
filelock==3.17.0
Flask==2.3.3
flask-cors==5.0.0
flask-swagger-ui==4.11.1
fsspec==2025.2.0
google==3.0.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.29.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.5
jiter==0.8.2
lxml==5.3.1
MarkupSafe==2.1.3
mlx==0.21.1
mlx-lm==0.20.6
numpy==2.2.3
ollama==0.4.7
openai==1.64.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
primp==0.14.0
protobuf==5.29.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.2
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
urllib3==2.3.0
waybackpy==2.4.4
websockets==14.2
Werkzeug==2.3.7
